Wrapper for `fping` to rapidly ping the network to fill the ARP cache.
This is useful if the ARP cache expires: by pinging all addresses on an
interface any devices that respond will have their MAC address recorded
for descrypi to find.

Usage: python3 fping.py eth0

We rely on `fping` to do the actual pinging. This wrapper takes an interface
name and constructs a command line to perform the ping with the least
amount of traffic (AFAIK), as quickly as possible without requiring root.

Restricted to IPv4 for now.